Fix torrent/downloads page error when qbittorrent downloads large files
Use Long.parseLong() to parse size of a torrent
instead of Integer.parseInt().

+++ home-media-app/src/main/java/com/gorolykmaxim/homemediaapp/qbittorrent/QbittorrentFactory.java
@@ -41,7 +41,7 @@ public class QbittorrentFactory {
         return new Qbittorrent(
                 qbittorrent.get("hash"),
                 qbittorrent.get("name"),
-                Integer.parseInt(qbittorrent.get("size")),
+                Long.parseLong(qbittorrent.get("size")),
                 Double.parseDouble(qbittorrent.get("progress")),
                 qbittorrent.get("state"),
                 new DownloadSpeed(Long.parseLong(qbittorrent.get("dlspeed")), "second"),
